Reverse transactions before adding them to the log
The new transactions which are returned when polling for updates are
received in reverse chronological order from the core. Iterating through
them to add them to the current list of transactions reverses this
order. Since the API does not support a parameter to change the sort
order the new transactions need to be reversed before they are added to
the log.

OHM-460