OHM-838 Fix transaction re-run permission error
The app currently processes the user's permissions based on the order of
the user groups from the API. What this means is that if a user belongs to
two or more user groups and the required group is not last in the list,
the user would have no permission to re-run transactions.

This is a UI logic related issue where the required checkbox for re-running
a transaction is not being displayed when it needs to be.

I have added a fix to correctly process the permissions related to
a channel and the logged in user.

Related Tickets:
 - OHM-834